// src/api/generated/models/resetPasswordRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Rukun API
 * Standard API contract for Core R.
 * OpenAPI spec version: 1.0.0
 */

export interface ResetPasswordRequest {
  token: string;
  email: string;
  /** @minLength 12 */
  password: string;
  password_confirmation: string;
}